completion for others
