+++ cmd/kaf/config.go
@@ -48,14 +48,6 @@ var configCurrentContext = &cobra.Command{
 	},
 }
 
-func validConfigArgs(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
-	clusterList := make([]string, 0, len(cfg.Clusters))
-	for _, cluster := range cfg.Clusters {
-		clusterList = append(clusterList, cluster.Name)
-	}
-	return clusterList, cobra.ShellCompDirectiveNoFileComp
-}
-
 var configUseCmd = &cobra.Command{
 	Use:               "use-cluster [NAME]",
 	Short:             "Sets the current cluster in the configuration",
