+++ cmd/kaf/consume.go
@@ -80,10 +80,11 @@ const (
 )
 
 var consumeCmd = &cobra.Command{
-	Use:    "consume TOPIC",
-	Short:  "Consume messages",
-	Args:   cobra.ExactArgs(1),
-	PreRun: setupProtoDescriptorRegistry,
+	Use:               "consume TOPIC",
+	Short:             "Consume messages",
+	Args:              cobra.ExactArgs(1),
+	ValidArgsFunction: validTopicArgs,
+	PreRun:            setupProtoDescriptorRegistry,
 	Run: func(cmd *cobra.Command, args []string) {
 		var offset int64
 		switch offsetFlag {
